separate misc module
