+++ rpc/types/src/lib.rs
@@ -104,28 +104,20 @@
 // TODO: remove me after finishing impl
 #![allow(dead_code)]
 
-//---------------------------------------------------------------------------------------------------- Use
-mod binary_string;
+//---------------------------------------------------------------------------------------------------- Mod
 mod constants;
 mod defaults;
 mod free;
 mod macros;
 
-pub use binary_string::BinaryString;
+pub mod base;
+pub mod bin;
+pub mod json;
+pub mod misc;
+pub mod other;
+
 pub use constants::{
     CORE_RPC_STATUS_BUSY, CORE_RPC_STATUS_NOT_MINING, CORE_RPC_STATUS_OK,
     CORE_RPC_STATUS_PAYMENT_REQUIRED, CORE_RPC_STATUS_UNKNOWN, CORE_RPC_VERSION,
     CORE_RPC_VERSION_MAJOR, CORE_RPC_VERSION_MINOR,
 };
-
-pub mod base;
-#[cfg(feature = "bin")]
-#[cfg_attr(docsrs, doc(cfg(feature = "bin")))]
-pub mod bin;
-#[cfg(feature = "json")]
-#[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-pub mod json;
-pub mod misc;
-#[cfg(feature = "other")]
-#[cfg_attr(docsrs, doc(cfg(feature = "other")))]
-pub mod other;
