+++ rpc/types/src/misc/misc.rs
@@ -4,12 +4,6 @@
 //! For example, [`crate::json::GetConnectionsResponse`] contains
 //! the [`crate::misc::ConnectionInfo`] struct defined here.
 
-//---------------------------------------------------------------------------------------------------- Lints
-#![allow(
-    missing_docs, // Docs are at: <https://www.getmonero.org/resources/developer-guides/daemon-rpc.html>
-    clippy::struct_excessive_bools, // hey man, tell that to the people who wrote `monerod`
-)]
-
 //---------------------------------------------------------------------------------------------------- Import
 use std::fmt::Display;
 
@@ -539,308 +533,6 @@ define_struct_and_impl_epee! {
     }
 }
 
-//---------------------------------------------------------------------------------------------------- TODO
-// TODO - weird types.
-
-#[doc = monero_definition_link!(
-    cc73fe71162d564ffda8e549b79a350bca53c454,
-    "rpc/core_rpc_server_commands_defs.h",
-    210..=221
-)]
-/// Used in [`crate::bin::GetBlocksResponse`].
-#[derive(Clone, Debug, Default, PartialEq, Eq, PartialOrd, Ord, Hash)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub struct BlockCompleteEntry {
-    pub pruned: bool,
-    pub block: String,
-    pub block_weight: u64,
-    pub txs: Vec<TxBlobEntry>,
-}
-
-// TODO: custom epee
-// <https://github.com/monero-project/monero/blob/cc73fe71162d564ffda8e549b79a350bca53c454/src/cryptonote_protocol/cryptonote_protocol_defs.h#L138-L163>
-#[cfg(feature = "epee")]
-epee_object! {
-    BlockCompleteEntry,
-    pruned: bool,
-    block: String,
-    block_weight: u64,
-    txs: Vec<TxBlobEntry>,
-}
-
-/// Used in [`crate::bin::GetBlocksResponse`].
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[repr(u8)]
-pub enum PoolInfoExtent {
-    None = 0,
-    Incremental = 1,
-    Full = 2,
-}
-
-// <https://github.com/monero-project/monero/blob/cc73fe71162d564ffda8e549b79a350bca53c454/src/cryptonote_protocol/cryptonote_protocol_defs.h#L138-L163>
-#[cfg(feature = "epee")]
-impl EpeeValue for PoolInfoExtent {
-    const MARKER: Marker = <String as EpeeValue>::MARKER;
-
-    fn read<B: Buf>(r: &mut B, marker: &Marker) -> cuprate_epee_encoding::Result<Self> {
-        todo!()
-    }
-
-    fn should_write(&self) -> bool {
-        todo!()
-    }
-
-    fn epee_default_value() -> Option<Self> {
-        todo!()
-    }
-
-    fn write<B: BufMut>(self, w: &mut B) -> cuprate_epee_encoding::Result<()> {
-        todo!()
-    }
-}
-
-#[doc = monero_definition_link!(
-    cc73fe71162d564ffda8e549b79a350bca53c454,
-    "rpc/core_rpc_server_commands_defs.h",
-    389..=428
-)]
-/// Used in [`crate::other::GetTransactionsResponse`].
-#[derive(Clone, Debug, Default, PartialEq, Eq, PartialOrd, Ord, Hash)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub struct TxEntry {
-    pub as_hex: String,
-    pub as_json: String,
-    pub block_height: u64,
-    pub block_timestamp: u64,
-    pub confirmations: u64,
-    pub double_spend_seen: bool,
-    pub in_pool: bool,
-    pub output_indices: Vec<u64>,
-    pub prunable_as_hex: String,
-    pub prunable_hash: String,
-    pub pruned_as_hex: String,
-    pub received_timestamp: u64,
-    pub relayed: bool,
-    pub tx_hash: String,
-}
-
-// TODO: custom epee
-// <https://github.com/monero-project/monero/blob/cc73fe71162d564ffda8e549b79a350bca53c454/src/cryptonote_protocol/cryptonote_protocol_defs.h#L406-427>
-#[cfg(feature = "epee")]
-epee_object! {
-    TxEntry,
-    as_hex: String,
-    as_json: String, // TODO: should be its own struct
-    block_height: u64,
-    block_timestamp: u64,
-    confirmations: u64,
-    double_spend_seen: bool,
-    in_pool: bool,
-    output_indices: Vec<u64>,
-    prunable_as_hex: String,
-    prunable_hash: String,
-    pruned_as_hex: String,
-    received_timestamp: u64,
-    relayed: bool,
-    tx_hash: String,
-}
-
-//---------------------------------------------------------------------------------------------------- TODO
-/// Used in [`crate::other::IsKeyImageSpentResponse`].
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[repr(u8)]
-pub enum KeyImageSpentStatus {
-    Unspent = 0,
-    SpentInBlockchain = 1,
-    SpentInPool = 2,
-}
-
-#[cfg(feature = "epee")]
-impl EpeeValue for KeyImageSpentStatus {
-    const MARKER: Marker = <String as EpeeValue>::MARKER;
-
-    fn read<B: Buf>(r: &mut B, marker: &Marker) -> cuprate_epee_encoding::Result<Self> {
-        todo!()
-    }
-
-    fn should_write(&self) -> bool {
-        todo!()
-    }
-
-    fn epee_default_value() -> Option<Self> {
-        todo!()
-    }
-
-    fn write<B: BufMut>(self, w: &mut B) -> cuprate_epee_encoding::Result<()> {
-        todo!()
-    }
-}
-
-//---------------------------------------------------------------------------------------------------- Status
-/// RPC response status.
-///
-/// This type represents `monerod`'s frequently appearing string field, `status`.
-///
-/// This field appears within RPC [JSON response](crate::json) types.
-///
-/// Reference: <https://github.com/monero-project/monero/blob/cc73fe71162d564ffda8e549b79a350bca53c454/src/rpc/core_rpc_server_commands_defs.h#L78-L81>.
-///
-/// ## Serialization and string formatting
-/// ```rust
-/// use cuprate_rpc_types::{
-///     misc::Status,
-///     CORE_RPC_STATUS_BUSY, CORE_RPC_STATUS_NOT_MINING, CORE_RPC_STATUS_OK,
-///     CORE_RPC_STATUS_PAYMENT_REQUIRED, CORE_RPC_STATUS_UNKNOWN
-/// };
-/// use serde_json::to_string;
-///
-/// let unknown = Status::Unknown;
-///
-/// assert_eq!(to_string(&Status::Ok).unwrap(),              r#""OK""#);
-/// assert_eq!(to_string(&Status::Busy).unwrap(),            r#""BUSY""#);
-/// assert_eq!(to_string(&Status::NotMining).unwrap(),       r#""NOT MINING""#);
-/// assert_eq!(to_string(&Status::PaymentRequired).unwrap(), r#""PAYMENT REQUIRED""#);
-/// assert_eq!(to_string(&unknown).unwrap(),                 r#""UNKNOWN""#);
-///
-/// assert_eq!(Status::Ok.as_ref(),              CORE_RPC_STATUS_OK);
-/// assert_eq!(Status::Busy.as_ref(),            CORE_RPC_STATUS_BUSY);
-/// assert_eq!(Status::NotMining.as_ref(),       CORE_RPC_STATUS_NOT_MINING);
-/// assert_eq!(Status::PaymentRequired.as_ref(), CORE_RPC_STATUS_PAYMENT_REQUIRED);
-/// assert_eq!(unknown.as_ref(),                 CORE_RPC_STATUS_UNKNOWN);
-///
-/// assert_eq!(format!("{}", Status::Ok),              CORE_RPC_STATUS_OK);
-/// assert_eq!(format!("{}", Status::Busy),            CORE_RPC_STATUS_BUSY);
-/// assert_eq!(format!("{}", Status::NotMining),       CORE_RPC_STATUS_NOT_MINING);
-/// assert_eq!(format!("{}", Status::PaymentRequired), CORE_RPC_STATUS_PAYMENT_REQUIRED);
-/// assert_eq!(format!("{}", unknown),                 CORE_RPC_STATUS_UNKNOWN);
-///
-/// assert_eq!(format!("{:?}", Status::Ok),              "Ok");
-/// assert_eq!(format!("{:?}", Status::Busy),            "Busy");
-/// assert_eq!(format!("{:?}", Status::NotMining),       "NotMining");
-/// assert_eq!(format!("{:?}", Status::PaymentRequired), "PaymentRequired");
-/// assert_eq!(format!("{:?}", unknown),                 "Unknown");
-/// ```
-#[derive(Copy, Clone, Debug, Default, PartialEq, Eq, PartialOrd, Ord, Hash)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub enum Status {
-    // FIXME:
-    // `#[serde(rename = "")]` only takes raw string literals?
-    // We have to re-type the constants here...
-    /// Successful RPC response, everything is OK; [`CORE_RPC_STATUS_OK`].
-    #[cfg_attr(feature = "serde", serde(rename = "OK"))]
-    #[default]
-    Ok,
-
-    /// The daemon is busy, try later; [`CORE_RPC_STATUS_BUSY`].
-    #[cfg_attr(feature = "serde", serde(rename = "BUSY"))]
-    Busy,
-
-    /// The daemon is not mining; [`CORE_RPC_STATUS_NOT_MINING`].
-    #[cfg_attr(feature = "serde", serde(rename = "NOT MINING"))]
-    NotMining,
-
-    /// Payment is required for RPC; [`CORE_RPC_STATUS_PAYMENT_REQUIRED`].
-    #[cfg_attr(feature = "serde", serde(rename = "PAYMENT REQUIRED"))]
-    PaymentRequired,
-
-    /// Some unknown other string; [`CORE_RPC_STATUS_UNKNOWN`].
-    ///
-    /// This exists to act as a catch-all if `monerod` adds
-    /// a string and a Cuprate node hasn't updated yet.
-    ///
-    /// The reason this isn't `Unknown(String)` is because that
-    /// disallows [`Status`] to be [`Copy`], and thus other types
-    /// that contain it.
-    #[cfg_attr(feature = "serde", serde(other))]
-    #[cfg_attr(feature = "serde", serde(rename = "UNKNOWN"))]
-    Unknown,
-}
-
-impl From<String> for Status {
-    fn from(s: String) -> Self {
-        match s.as_str() {
-            CORE_RPC_STATUS_OK => Self::Ok,
-            CORE_RPC_STATUS_BUSY => Self::Busy,
-            CORE_RPC_STATUS_NOT_MINING => Self::NotMining,
-            CORE_RPC_STATUS_PAYMENT_REQUIRED => Self::PaymentRequired,
-            _ => Self::Unknown,
-        }
-    }
-}
-
-impl AsRef<str> for Status {
-    fn as_ref(&self) -> &str {
-        match self {
-            Self::Ok => CORE_RPC_STATUS_OK,
-            Self::Busy => CORE_RPC_STATUS_BUSY,
-            Self::NotMining => CORE_RPC_STATUS_NOT_MINING,
-            Self::PaymentRequired => CORE_RPC_STATUS_PAYMENT_REQUIRED,
-            Self::Unknown => CORE_RPC_STATUS_UNKNOWN,
-        }
-    }
-}
-
-impl Display for Status {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(self.as_ref())
-    }
-}
-
-// [`Status`] is essentially a [`String`] when it comes to
-// (de)serialization, except when writing we usually have
-// access to a `&'static str` and don't need to allocate.
-//
-// See below for more impl info:
-// <https://github.com/Cuprate/cuprate/blob/bef2a2cbd4e1194991751d1fbc96603cba8c7a51/net/epee-encoding/src/value.rs#L366-L392>.
-#[cfg(feature = "epee")]
-impl EpeeValue for Status {
-    const MARKER: Marker = <String as EpeeValue>::MARKER;
-
-    fn read<B: Buf>(r: &mut B, marker: &Marker) -> cuprate_epee_encoding::Result<Self> {
-        let string = <String as EpeeValue>::read(r, marker)?;
-        Ok(Self::from(string))
-    }
-
-    fn should_write(&self) -> bool {
-        true
-    }
-
-    fn epee_default_value() -> Option<Self> {
-        // <https://github.com/Cuprate/cuprate/pull/147#discussion_r1654992559>
-        Some(Self::Unknown)
-    }
-
-    fn write<B: BufMut>(self, w: &mut B) -> cuprate_epee_encoding::Result<()> {
-        cuprate_epee_encoding::write_bytes(self.as_ref(), w)
-    }
-}
-
 //---------------------------------------------------------------------------------------------------- Tests
 #[cfg(test)]
-mod test {
-    use super::*;
-
-    // Test epee (de)serialization works.
-    #[test]
-    #[cfg(feature = "epee")]
-    fn epee() {
-        for status in [
-            Status::Ok,
-            Status::Busy,
-            Status::NotMining,
-            Status::PaymentRequired,
-            Status::Unknown,
-        ] {
-            let mut buf = vec![];
-
-            <Status as EpeeValue>::write(status, &mut buf).unwrap();
-            let status2 =
-                <Status as EpeeValue>::read(&mut buf.as_slice(), &<Status as EpeeValue>::MARKER)
-                    .unwrap();
-
-            assert_eq!(status, status2);
-        }
-    }
-}
+mod test {}
